requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.50.0
django-cors-headers==3.12.0
django-heroku==0.3.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.9
whitenoise==6.1.0
